+++ examples/hello-world/src/main.rs
@@ -179,7 +179,7 @@ impl DXSample for Sample {
 
         let rtv_descriptor_size =
             self.device
-                .get_descriptor_handle_increment_size(DescriptorHeapType::Rtv) as usize;
+                .get_descriptor_handle_increment_size(DescriptorHeapType::Rtv);
         let rtv_handle = rtv_heap.get_cpu_descriptor_handle_for_heap_start();
 
         let render_targets: [Resource; FRAME_COUNT as usize] = std::array::from_fn(|i| {
