+++ oxidx/src/device.rs
@@ -297,7 +297,7 @@ pub trait IDevice: HasInterface<Raw: Interface> {
     /// Gets the size of the handle increment for the given type of descriptor heap. This value is typically used to increment a handle into a descriptor array by the correct amount.
     ///
     /// For more information: [`ID3D12Device::GetDescriptorHandleIncrementSize method`](https://learn.microsoft.com/en-us/windows/win32/api/d3d12/nf-d3d12-id3d12device-getdescriptorhandleincrementsize)
-    fn get_descriptor_handle_increment_size(&self, r#type: DescriptorHeapType) -> u32;
+    fn get_descriptor_handle_increment_size(&self, r#type: DescriptorHeapType) -> usize;
 
     /// Gets the reason that the device was removed, or [`Result::Ok`] if the device isn't removed.
     /// To be called back when a device is removed, consider using [`IFence::set_event_on_completion`] with a value of [`u64::MAX`].
@@ -916,9 +916,9 @@ impl_trait! {
         }
     }
 
-    fn get_descriptor_handle_increment_size(&self, r#type: DescriptorHeapType) -> u32 {
+    fn get_descriptor_handle_increment_size(&self, r#type: DescriptorHeapType) -> usize {
         unsafe {
-            self.0.GetDescriptorHandleIncrementSize(r#type.as_raw())
+            self.0.GetDescriptorHandleIncrementSize(r#type.as_raw()) as usize
         }
     }
